Rename books to bookSlice
